requirements.txt
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-markdown-deux==1.0.5
django-pagedown==2.1.2
djangorestframework==3.12.1
Markdown==3.3.2
markdown2==2.3.9
Pillow==7.2.0
